refactor(folder-structure): add folders to components
