+++ client/src/components/Form/Form.jsx
@@ -8,15 +8,19 @@ import {
   IconButton,
 } from "@mui/material";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
-import { useAuth, useSnackbar } from "../hooks";
-import GoogleLoginAuth from "./GoogleLoginAuth";
-import GoogleSheetsButton from "./GoogleSheetsButton";
+import { useAuth, useSnackbar } from "../../hooks";
+
+import {
+  ErrorMessage,
+  GoogleLoginAuth,
+  GoogleSheetsButton,
+  FormButton,
+} from "../../components";
 import AddIcon from "@mui/icons-material/Add";
 import RemoveCircleOutlineIcon from "@mui/icons-material/RemoveCircleOutline";
 import axios from "axios";
-import ErrorMessage from "./ErrorMessage";
+
 import HighlightOffIcon from "@mui/icons-material/HighlightOff";
-import FormButton from "./FormButton";
 
 const Form = () => {
   const { user } = useAuth();
