+++ client/src/components/GoogleLoginAuth/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { GoogleLogin } from "@react-oauth/google";
 import { jwtDecode } from "jwt-decode";
 import { useNavigate } from "react-router";
-import { useAuth, useSnackbar } from "../hooks";
+import { useAuth, useSnackbar } from "../../hooks";
 
 const GoogleLoginAuth = () => {
   const { setUser } = useAuth();
