+++ client/src/components/index.js
@@ -7,9 +7,10 @@ export { default as SignIn } from "./SignIn/SignIn";
 export { default as SignInCard } from "./SignIn/SignInCard";
 
 //Components
-export { default as ErrorMessage } from "./ErrorMessage";
-export { default as Form } from "./Form";
-export { default as GoogleLoginAuth } from "./GoogleLoginAuth";
-export { default as GoogleSheetsButton } from "./Form";
+export { default as ErrorMessage } from "./ErrorMessage/index";
+export { default as Form } from "./Form/Form";
+export { default as FormButton } from "./Form/FormButton";
+export { default as GoogleLoginAuth } from "./GoogleLoginAuth/index";
+export { default as GoogleSheetsButton } from "./GoogleSheetsButton/index";
 export { default as TitleWrapper } from "./TitleWrapper";
 export { default as PositionedSnackbar } from "./PositionedSnackbar/index";
